Removed and reorganized code
